Убрал нереализованная клавиатуру, вынес new DefaultBotOptions в отделный абстрактный класс.
