+++ tibbo-calculator/src/main/java/com/tibbo/MainBot.java
@@ -1,0 +1,19 @@
+package com.tibbo;
+
+import org.telegram.telegrambots.ApiContextInitializer;
+import org.telegram.telegrambots.meta.TelegramBotsApi;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+public class MainBot{
+
+    public static void main(String[] args) {
+        ApiContextInitializer.init(); // Инициализируем апи
+        TelegramBotsApi botapi = new TelegramBotsApi();
+        try {
+            botapi.registerBot(new TelegramBot());
+        } catch (TelegramApiException e) {
+            e.printStackTrace();
+        }
+
+    }
+}
