+++ tibbo-calculator/src/main/java/com/tibbo/TelegramBot.java
@@ -1,54 +1,65 @@
 package com.tibbo;
 
-import org.telegram.telegrambots.ApiContextInitializer;
-import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardMarkup;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.KeyboardButton;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.KeyboardRow;
+import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
+import java.net.InetSocketAddress;
+import java.net.Socket;
 
-public class TelegramBot{
+public class TelegramBot extends BotOptions {
+    private static final String USERNAME = "Lourence_bot";
+    private static final String TOKEN = "914836872:AAF6sEyC4igukZ860X7bZWorhAjs0AE8VUQ";
+    private Socket socket = new Socket();
 
-    public static void main(String[] args) {
-        ApiContextInitializer.init(); // Инициализируем апи
-        TelegramBotsApi botapi = new TelegramBotsApi();
+    private static int getPorts(){
+        return 5555;
+    }
+    private String getMsg(String msg){
         try {
-            botapi.registerBot(new BotOptions());
-        } catch (TelegramApiException e) {
+            String host = "localhost";
+            socket.connect(new InetSocketAddress(host, getPorts()));
+            DataOutputStream outStream= new DataOutputStream(socket.getOutputStream());
+            DataInputStream inputStream = new DataInputStream(socket.getInputStream());
+            outStream.writeUTF(msg);
+            outStream.flush();
+            return inputStream.readUTF();
+        }catch (IOException e) {
+            e.printStackTrace();
+        }
+        return msg;
+    }
+
+    public void onUpdateReceived(Update update) {
+        long chat_id = update.getMessage().getChatId();
+
+        SendMessage sendMessage = new SendMessage().setChatId(chat_id);
+        String text = update.getMessage().getText();
+
+        try {
+
+            sendMessage.setText("Вы ввели: " + text);
+            execute(sendMessage);
+
+            sendMessage.setText("Результат: " + getMsg(text));
+            execute(sendMessage);
+
+
+        }catch (TelegramApiException e){
             e.printStackTrace();
         }
 
     }
 
-    private synchronized void setButtons(SendMessage sendMessage) { // клавиатура еще не реализована
-        // Создаем клавиуатуру
-        ReplyKeyboardMarkup replyKeyboardMarkup = new ReplyKeyboardMarkup();
-        sendMessage.setReplyMarkup(replyKeyboardMarkup);
-        replyKeyboardMarkup.setSelective(true);
-        replyKeyboardMarkup.setResizeKeyboard(true);
-        replyKeyboardMarkup.setOneTimeKeyboard(false);
+    public String getBotUsername() {
+        return USERNAME;
+    }
 
-        // Создаем список строк клавиатуры
-        List<KeyboardRow> keyboard = new ArrayList<>();
-
-        // Первая строчка клавиатуры
-        KeyboardRow keyboardFirstRow = new KeyboardRow();
-        // Добавляем кнопки в первую строчку клавиатуры
-        keyboardFirstRow.add(new KeyboardButton("Button"));
-
-        // Вторая строчка клавиатуры
-        KeyboardRow keyboardSecondRow = new KeyboardRow();
-        // Добавляем кнопки во вторую строчку клавиатуры
-        keyboardSecondRow.add(new KeyboardButton("Button"));
-
-        // Добавляем все строчки клавиатуры в список
-        keyboard.add(keyboardFirstRow);
-        keyboard.add(keyboardSecondRow);
-        // и устанваливаем этот список нашей клавиатуре
-        replyKeyboardMarkup.setKeyboard(keyboard);
+    @Override
+    public String getBotToken() {
+        return TOKEN;
     }
 }
